==> src/app/notebooks/output/2.0_data_preparation_output.ipynb <==
# Injected parameters
train_dim = 0.7
val_test_dim = 0.15


import sys
sys.path.append('../../../')

import pandas as pd
import numpy as np
import streamlit as st

from sklearn.model_selection import train_test_split
from catboost import Pool, CatBoostClassifier
from sklearn.metrics import roc_auc_score

pd.set_option("display.max_rows",None)
pd.set_option("display.max_columns",None)

def partitioning(df,train_dim, val_test_dim,target):
    """
    
    """
    val_test_dim_edit = train_dim + val_test_dim
    train, val, test = np.split(df.sample(frac=1,random_state=2), [int(train_dim*len(df)), int(val_test_dim_edit*len(df))])
    st.write("Train shape: ",train.shape)
    st.write("Train %: \n",train[target].value_counts(normalize=True))
    st.write("Val %: \n",val[target].value_counts(normalize=True))
    st.write("Test %: \n",test[target].value_counts(normalize=True))
    return train, val, test

def auc_score(y_trues, y_preds):
    """
    
    """
    for i, y_pred in enumerate(y_preds):
        y_true = y_trues[i]
        auc = roc_auc_score(y_true, y_pred)
    return auc

def adversarial_validation(val,test,drop_cols):
    """
    
    """
    aval = val.copy()
    atest = test.copy()
    # 1 - define target
    aval['y'] = 1.0
    atest['y'] = 0.0
    # 2 - create dataframe
    ad = aval.append(atest).sample(frac=1,random_state=2)
    # 3 - drop unuseful columns
    c_drop = drop_cols
    ad = ad.drop(columns=c_drop)
    # 4 - define format and imputation
    for i in ad:
        if ad[i].dtypes != 'object':
            ad[i] = ad[i].astype(float)
            ad[i] = ad[i].fillna(ad[i].median())
        else:
            ad[i] = ad[i].fillna('missing')
            ad[i] = ad[i].astype(str)
    # 5 - model preparation
    y = ad['y'].values
    X = ad.drop(columns=['y']).values
    categorical_features_indices = np.where(ad.dtypes != np.float64)[0]
    # 6 - train test split
    adv_X_train, adv_X_test, adv_y_train, adv_y_test = train_test_split(X, y , test_size = 0.30 , random_state = 2)
    train_data = Pool(data=adv_X_train,label=adv_y_train,cat_features=categorical_features_indices)
    test_data = Pool(data=adv_X_test,label=adv_y_test,cat_features=categorical_features_indices)
    # 7 - model training
    params = {'iterations': 1000,'eval_metric': 'AUC','od_type': 'Iter','od_wait': 50}
    model = CatBoostClassifier(**params)
    _ = model.fit(train_data, eval_set=test_data, plot=False, verbose=False)
    # 8 - model evaluation
    auc = auc_score([test_data.get_label()],[model.predict_proba(test_data)[:,1]])
    if auc <= 0.6:
        return st.write("No distribution shift, OK! AUC is: ",auc)
    else:
        return st.write("Check features importance (to be added) and rerun. AUC is: ",auc)

def class_imbalance(train,target):
    """
    
    """
    df_under = train[train[target]==train[target].value_counts(normalize=False).index[1]]
    df_over = train[train[target]==train[target].value_counts(normalize=False).index[0]]
    df_over = df_over.sample(frac=df_under.shape[0]/df_over.shape[0],random_state=2)
    final_train = df_over.append(df_under).sample(frac=1,random_state=2)
    st.write(final_train[target].value_counts(normalize=False))
    st.write("Train dataset shape: ",final_train.shape)
    return final_train

# 1.0 Data retrieval

df = pd.read_pickle("data_lake/output/df_under.pkl")

df.sample(2)

df.shape

# 2.0 Data Partitioning

df['isFraud'].value_counts()

## 2.1 Train, validation, test split

train_dim = 0.7
val_test_dim = 0.15

train, val, test = partitioning(df,train_dim, val_test_dim,'isFraud')

## 2.2 Adversarial validation between val and test set

drop_cols = ['TransactionID','customer_id']

adversarial_validation(val,test,drop_cols)

## 2.3 Class imbalance

final_train = class_imbalance(train,'isFraud')

# 3.0 Output

final_train.to_pickle('app/datalake/train.pkl')
val.to_pickle('app/datalake/val.pkl')
